Refactor and Automapper

+++ EmployeeMonthlyPayslipApp.Models/Models/EmployeeDetails.cs
@@ -15,7 +15,7 @@ namespace EmployeeMonthlyPayslipApp.Models.Models
     {
         public string FirstName { get; set; }
         public string LastName { get; set; }
-        public decimal AnnualSalary { get; set; }
+        public int AnnualSalary { get; set; }
         public string TaxPeriod { get; set; }
         public string SuperPercentage { get; set; }
     }
